requirements.txt
# OmniSorter Standalone - Minimal Dependencies
# 最小限のパッケージ構成
# Python 3.12 対応版

streamlit>=1.30.0
pandas>=2.0.3
numpy>=1.26.0
plotly>=5.17.0
PyYAML>=6.0.1
